+++ elems/tlorenz.cpp
@@ -56,7 +56,7 @@ double TLorenz::f() noexcept
 }
 
 
-int TLorenz::miso_startLoop( int /*acnx*/, int /*acny*/ )
+int TLorenz::miso_startLoop( long /*acnx*/, long /*acny*/ )
 {
   x = (double)x_0; y = (double)y_0 ; z = (double)z_0;
   out0 = x; // ????
