+++ elems/trossler.cpp
@@ -54,7 +54,7 @@ double TRossler::f() noexcept
   return x;
 }
 
-int TRossler::miso_startLoop( int /*acnx*/, int /*acny*/ )
+int TRossler::miso_startLoop( long /*acnx*/, long /*acny*/ )
 {
   x = (double)x_0; y = (double)y_0 ; z = (double)z_0;
   out0 = x;
